backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
supabase==2.9.1

# External data sources
httpx==0.27.2
yfinance==0.2.43

# Auth
PyJWT==2.9.0
python-jose[cryptography]==3.3.0

# Config
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx  # also used for test client

# Utilities
python-dateutil==2.9.0
